requirements.txt
requests
setuptools
qrcode
loguru
Pillow
retry
tinydb
ntplib
playsound==1.2.2
retrying
tabulate
wheel
inquirer
logtail
urllib3
sentry-sdk>=1.40.0
py-machineid

